move inputForRun fetcher into build wiring and optimze imports
