+++ src/main/java/org/icgc_argo/workflow/search/controller/impl/RunsApiController.java
@@ -25,7 +25,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.icgc_argo.workflow.search.controller.RunsApi;
 import org.icgc_argo.workflow.search.model.common.RunId;
 import org.icgc_argo.workflow.search.model.common.RunRequest;
-import org.icgc_argo.workflow.search.model.wes.*;
+import org.icgc_argo.workflow.search.model.wes.RunListResponse;
+import org.icgc_argo.workflow.search.model.wes.RunResponse;
+import org.icgc_argo.workflow.search.model.wes.RunStatus;
+import org.icgc_argo.workflow.search.model.wes.ServiceInfo;
 import org.icgc_argo.workflow.search.service.wes.WesRunService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
