+++ src/main/java/org/icgc_argo/workflow/search/graphql/fetchers/RunDataFetchers.java
@@ -18,11 +18,15 @@
 
 package org.icgc_argo.workflow.search.graphql.fetchers;
 
+import static com.google.common.base.Strings.isNullOrEmpty;
 import static java.util.stream.Collectors.toUnmodifiableList;
+import static org.icgc_argo.workflow.search.model.SearchFields.ANALYSIS_ID;
+import static org.icgc_argo.workflow.search.util.Converter.asImmutableMap;
 import static org.icgc_argo.workflow.search.util.JacksonUtils.convertValue;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
@@ -30,12 +34,14 @@ import lombok.val;
 import org.icgc_argo.workflow.search.graphql.AsyncDataFetcher;
 import org.icgc_argo.workflow.search.model.common.Run;
 import org.icgc_argo.workflow.search.model.common.Task;
-import org.icgc_argo.workflow.search.model.graphql.*;
+import org.icgc_argo.workflow.search.model.graphql.AggregationResult;
+import org.icgc_argo.workflow.search.model.graphql.Analysis;
 import org.icgc_argo.workflow.search.model.graphql.SearchResult;
 import org.icgc_argo.workflow.search.model.graphql.Sort;
 import org.icgc_argo.workflow.search.service.graphql.RunService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import reactor.core.publisher.Mono;
 
 @Slf4j
 @Component
@@ -87,10 +93,30 @@ public class RunDataFetchers {
     };
   }
 
-  public AsyncDataFetcher<Run> getNestedRunDataFetcher() {
+  public AsyncDataFetcher<Run> getNestedRunInTaskDataFetcher() {
     return environment -> {
       val task = (Task) environment.getSource();
       return runService.getRunByRunId(task.getRunId());
+    };
+  }
+
+  public AsyncDataFetcher<List<Run>> getNestedRunInAnalysisDataFetcher() {
+    return environment -> {
+      val analysis = (Analysis) environment.getSource();
+      val analysisId = analysis.getAnalysisId();
+
+      ImmutableMap<String, Object> filter = asImmutableMap(environment.getArgument("filter"));
+      val filerAnalysisId = filter.getOrDefault(ANALYSIS_ID, analysisId);
+
+      // short circuit here since can't find runs for invalid analysisId
+      if (isNullOrEmpty(analysisId) || !analysisId.equals(filerAnalysisId)) {
+        return Mono.empty();
+      }
+
+      Map<String, Object> mergedFilter = new HashMap<>(filter);
+      mergedFilter.put(ANALYSIS_ID, analysisId);
+
+      return runService.getRuns(mergedFilter, null);
     };
   }
 }
