+++ src/main/resources/schema.graphqls
@@ -93,7 +93,7 @@ directive @fetch(from : String!) on FIELD_DEFINITION
 
 type Analysis @key(fields: "analysisId") @extends {
     analysisId: ID! @external
-    inputForRuns(filter: RunsFilter): [Run] @fetch(from: "inputForRuns")
+    inputForRuns(filter: RunsFilter): [Run]
 }
 
 type Workflow @key(fields: "runId") @extends {
